fix(customer):
- fix order service's pom.xml file.
- fix order service's application.yml file.
- fix order service's DB migration file.
- fix order service's mapping logic.

+++ services/order-service/src/main/resources/db/migration/V1__create_order_table.sql
@@ -3,7 +3,7 @@ create table if not exists orders.orders(
     customer_id uuid not null ,
     currency_from varchar(5) not null ,
     currency_to varchar(5) not null ,
-    amount bigint not null ,
-    rate bigint not null ,
+    amount decimal not null ,
+    rate decimal not null ,
     status varchar(25) not null
 )
